initial working separation of book and flow files
