+++ src/makeBook/canSplit.js
