+++ src/makeBook/ensureImageLoaded.js
