+++ src/makeBook/ensureImageLoaded.test.js
