+++ src/makeBook/estimateProgress.js
